# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 2Mart ERP API
  version: 0.1.0
  description: API Specs cho Hệ thống 2Mart ERP Mini
servers:
  - url: http://localhost:3000/api/v1
    description: Local server
...
paths:
...
  /products:
    get:
      tags:
        - Catalog
      operationId: searchProducts
      summary: Tìm kiếm sản phẩm
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductLookupDTO'

    post:
      tags:
        - Catalog
      operationId: createProduct
      summary: Tạo sản phẩm mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductLookupDTO'
...
components:
  schemas:
...
    ProductLookupDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        brandName:
          type: string
        categoryName:
          type: string
        retailPrice:
          type: number
        status:
          type: string
          enum: [ACTIVE, INACTIVE, DISCONTINUED, COMING_SOON]
        imageUrl:
          type: string
          nullable: true
    CreateProductDTO:
      type: object
      required:
        - sku
        - name
        - defaultUnitId
      properties:
        sku:
          type: string
        name:
          type: string
        categoryId:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        defaultUnitId:
          type: string
          format: uuid
        barcodes:
          type: array
          items:
            type: string
        retailPrice:
          type: number
...